Move light-tests.js tests to test.js file
